+++ bios_pagos/src/main/java/uy/edu/bios/ejemplos/biospagos/servicios/ServicioEnviosDinero.java
@@ -1,5 +1,6 @@
 package uy.edu.bios.ejemplos.biospagos.servicios;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -7,16 +8,16 @@ import org.springframework.stereotype.Service;
 
 import uy.edu.bios.ejemplos.biospagos.dominio.Cliente;
 import uy.edu.bios.ejemplos.biospagos.dominio.EnvioDinero;
-import uy.edu.bios.ejemplos.biospagos.repositorios.RepositorioEnvioDinero;
-
-
+import uy.edu.bios.ejemplos.biospagos.excepciones.ExcepcionBiosPagos;
+import uy.edu.bios.ejemplos.biospagos.excepciones.ExcepcionNoExiste;
+import uy.edu.bios.ejemplos.biospagos.repositorios.IRepositorioEnviosDinero;
 
 @Service
 public class ServicioEnviosDinero implements IServicioEnviosDinero {
 
-    private final RepositorioEnvioDinero repositorioEnvioDinero;
+    private final IRepositorioEnviosDinero repositorioEnvioDinero;
 
-    public ServicioEnviosDinero(RepositorioEnvioDinero repositorioEnvioDinero) {
+    public ServicioEnviosDinero(IRepositorioEnviosDinero repositorioEnvioDinero) {
         this.repositorioEnvioDinero = repositorioEnvioDinero;
     }
 
@@ -31,12 +32,30 @@ public class ServicioEnviosDinero implements IServicioEnviosDinero {
     }
 
     @Override
-    public EnvioDinero guardar(EnvioDinero envio) {
+    public EnvioDinero guardar(EnvioDinero envio) throws ExcepcionBiosPagos {
+
+        if (envio.getId() == null) {
+            envio.setFechaHoraRegistro(LocalDateTime.now());
+        } else {
+            EnvioDinero envioAnterior = repositorioEnvioDinero.findById(envio.getId()).orElse(null);
+
+            if (envioAnterior == null) {
+                throw new ExcepcionNoExiste("No existe el envío indicado.");
+            }
+
+            envio.setFechaHoraRegistro(envioAnterior.getFechaHoraRegistro());
+        }
+
         return repositorioEnvioDinero.save(envio);
     }
 
     @Override
-    public void eliminar(Long id) {
+    public void eliminar(Long id) throws ExcepcionBiosPagos {
+
+        if (!repositorioEnvioDinero.existsById(id)) {
+            throw new ExcepcionNoExiste("No existe el envío indicado.");
+        }
+
         repositorioEnvioDinero.deleteById(id);
     }
